\name{as.surveybraid}
\alias{as.surveybraid}
\title{Creates object of class surveybraid.}
\usage{
  as.surveybraid(surveyor, braid)
}
\arguments{
  \item{surveyor}{surveyor object}

  \item{braid}{braid object}
}
\value{
  A list object of class \code{surveyorbraid}
}
\description{
  A surveybraid object consists of a surveyor and braid
  object.
}

