\name{as.surveybraidppt}
\alias{as.surveybraidppt}
\title{Creates object of class surveybraidppt.}
\usage{
  as.surveybraidppt(surveyor, braid)
}
\arguments{
  \item{surveyor}{surveyor object}

  \item{braid}{braid object}
}
\value{
  A list object of class \code{surveyorbraid}
}
\description{
  A surveybraidppt object consists of a surveyor and
  braidppt object.
}

