\docType{package}
\name{surveybraid-package}
\alias{surveybraid}
\alias{surveybraid-package}
\title{Tools for reporting on survey objects.}
\description{
  A report writing system for surveys, using the braid
  package.
}
\author{
  Andrie de Vries \email{andrie@pentalibra.com}
}
\seealso{
  \code{\link{as.surveybraid}}
}
\keyword{package}

